// src/app/modules/logistica/integracao/tms/models/logisticaIntegracoesTMSTIDNotaFiscal.ts
export interface ILogisticaIntegracoesTMSTIDNotaFiscal {
  ID_LOGI_SENI_INTE_TID_NOTA_FISC: number;
  NM_ARQU: string;
  LOCA_ARQU: string;
  NR_NOTA_FISC: number;
  DS_ERRO: string;
  IN_IMPO: string;
  DT_IMPO: string;
  IN_PAGI: string;
  ORDE_BY: string;
  ORDE_TYPE: string;
  PAGI: number;
  TT_REGI_PAGI: number;

}